flying-sam/requirements.txt
addict==2.4.0
alembic==1.8.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
bcrypt==4.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
docker==7.1.0
executing==2.2.0
fastapi==0.115.8
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groundingdino-py==0.4.0
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mobile_sam @ git+https://github.com/ChaoningZhang/MobileSAM.git@34bbbfdface3c18e5221aa7de6032d7220c6c6a1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
parso==0.8.4
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pycocotools==2.0.8
pydantic==2.8.2
pydantic-settings==2.7.1
pydantic_core==2.20.1
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
segment-anything==1.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
stack-data==0.6.3
starlette==0.45.3
supervision==0.6.0
sympy==1.13.1
timm==1.0.14
tokenizers==0.21.0
torch==2.6.0+cu118
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.13
yapf==0.43.0
zenml==0.74.0